requirements.txt
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
aiohttp==3.9.1
Pillow==10.1.0
apscheduler==3.10.4
python-dotenv==1.0.0
lxml==4.9.3
pandas==2.1.3
numpy==1.24.3
newspaper3k==0.2.8
googletrans==4.0.0rc1
httpx==0.25.2